refactor intp, fix bug: a2f
