+++ aiida_supercon/workflows/protocols/a2f.yaml
@@ -1,32 +1,21 @@
 default_inputs:
   clean_workdir: False
-  qfpoints_distance: 0.1
-  kfpoints_factor: 1
-  epw:
-      metadata:
-      options:
-        resources:
-          num_machines: 1
-        max_wallclock_seconds: 43200  # Twelve hours
-        withmpi: True
   a2f:
-    metadata:
-      options:
-        resources:
-          num_machines: 1
-        max_wallclock_seconds: 43200  # Twelve hours
-        withmpi: True
-    parameters:
-      INPUTEPW:
-        eliashberg: True
-        elph: True
-        ep_coupling: True
-        epwread: True
-        epwwrite: False
-        ephwrite: True
-        wannierize: False
-        muc: 0.13
-        mp_mesh_k: True
+    epw:
+      metadata:
+        options:
+          withmpi: True
+      parameters:
+        INPUTEPW:
+          eliashberg: True
+          elph: True
+          ep_coupling: True
+          epwread: True
+          epwwrite: False
+          ephwrite: True
+          wannierize: False
+          muc: 0.13
+          mp_mesh_k: True
 
 default_protocol: moderate
 protocols:
@@ -34,7 +23,5 @@ protocols:
     description: 'Protocol to perform a electron-phonon calculation at normal precision at moderate computational cost.'
   precise:
     description: 'Protocol to perform a electron-phonon calculation at high precision at higher computational cost.'
-    qfpoints_distance: 0.05
   fast:
     description: 'Protocol to perform a electron-phonon calculation at low precision at minimal computational cost for testing purposes.'
-    qfpoints_distance: 0.2
